requirements.txt
morph-data==0.2.0
plotly
langchain
langchain-openai